Add files via upload
The purpose of this research is to know the record management of imported goods
notificationat customs activities for the success of dwelling time in Directorate
General of Customs andExcise Type A Tanjung Priok. Where in the management of
records archive documents import notification of goods through several processes
starting from the process of recording, organization, until the dissemination process.
The method used in this research is qualitative with case study approach. Data
collection techniques are taken through the process of observation, interviews, and
document studies. The result of the research shows that the process of recording the
document of goods import notification through the process of creation and receipt of
import documents where import documents are created from the delivery of import
goods notification documents (PIB) to the Customs Office submitted through
Electronic Data Exchange (PDE) system until the determination of document path
only in 4 minutes. In the process of the organization, the import declaration
documents through 5 (five) stages, namely the distribution of documents, the process
of completion of customs administration of import documents, the storage of
completed documents processed, the rediscovery of documents, and the maintenance
of import documents, which at this stage dwelling time at the Port of Tanjung Priok
decreased in April 2018 to 3.42 days. Furthermore, the dissemination process at the
Customs and Excise agency itself can not be separated from a system called
Customs-Excise Intergrated System and Automation (CEISA), a system integration of
all services of the Directorate General of Customs and Excise to all public service
users so that all service users as users can access Customs-Excise Intergrated System
and Automation (CEISA) system from anywhere and anytime as long as the user is on
an internet connection.

Keywords: record management, imported goods notification documents, customs
clearance, dwelling time.